Resolved a Bug in 3rd Integration

+++ Helperland/HelperLand/HelperLand/ViewModels/BookServiceCombinedViewModel.cs
@@ -7,10 +7,10 @@
         public ScheduleAndPlanViewModel scheduleAndPlan { get; set; }
         public AddressViewModel address { get; set; }
 
-        public int startTime { get; set; }
-        public int totalamount { get; set; }
-        public int totalservicetime { get; set; }
-        public int extraservicetime { get; set; }
-        public int basichrs { get; set; }
+        public float startTime { get; set; }
+        public float totalamount { get; set; }
+        public float totalservicetime { get; set; }
+        public float extraservicetime { get; set; }
+        public float basichrs { get; set; }
     }
 }
